requirements.txt
asgiref==3.2.10
cffi==1.14.0
cryptography==2.9.2
Django==3.0.7
django-filter==2.3.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
ssh2-python==0.18.0.post1
